requirements.txt
google-adk
litellm
litellm[proxy]